+++ src/FreeElephants/PsrRouter/PathNormalization/Dummy.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FreeElephants\PsrRouter\PathNormalization;
+
+class Dummy implements PathNormalizerInterface
+{
+    public function normalizePath(string $path): string
+    {
+        return $path;
+    }
+}