+++ src/FreeElephants/PsrRouter/PathNormalization/PathNormalizerInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace FreeElephants\PsrRouter\PathNormalization;
+
+interface PathNormalizerInterface
+{
+    public function normalizePath(string $path): string;
+}