+++ src/FreeElephants/PsrRouter/PathNormalization/TrailingSlashTrimmer.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FreeElephants\PsrRouter\PathNormalization;
+
+class TrailingSlashTrimmer implements PathNormalizerInterface
+{
+    public function normalizePath(string $path): string
+    {
+        return rtrim($path, '/');
+    }
+}