+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Path normalizing for trailing slash
+
 ## [0.0.1] - 2025-06-11
 
 ### Added